Rename source and header files
