+++ src/pegr/gensys/Lua_Interf_Runtime.cpp
@@ -9,10 +9,10 @@
 #include <vector>
 #include <map>
 
-#include "pegr/debug/DebugMacros.hpp"
+#include "pegr/debug/Debug_Macros.hpp"
 #include "pegr/logger/Logger.hpp"
 #include "pegr/gensys/Compiler.hpp"
-#include "pegr/script/ScriptHelper.hpp"
+#include "pegr/script/Script_Helper.hpp"
 #include "pegr/gensys/Runtime.hpp"
 #include "pegr/gensys/Gensys.hpp"
