+++ src/pegr/main.cpp
@@ -6,9 +6,9 @@
 #include "pegr/gensys/Gensys.hpp"
 #include "pegr/scheduler/Lua_Interf.hpp"
 #include "pegr/script/Script.hpp"
-#include "pegr/script/ScriptHelper.hpp"
+#include "pegr/script/Script_Helper.hpp"
 #include "pegr/logger/Logger.hpp"
-#include "pegr/debug/DebugMacros.hpp"
+#include "pegr/debug/Debug_Macros.hpp"
 
 using namespace pegr;
