+++ src/pegr/script/Lua_Interf_Util.cpp
@@ -4,9 +4,9 @@
 #include <stdexcept>
 #include <cassert>
 
-#include "pegr/debug/DebugMacros.hpp"
+#include "pegr/debug/Debug_Macros.hpp"
 #include "pegr/script/Script.hpp"
-#include "pegr/script/ScriptHelper.hpp"
+#include "pegr/script/Script_Helper.hpp"
 #include "pegr/logger/Logger.hpp"
 
 namespace pegr {
