+++ src/pegr/test/Script_Helper_Test.cpp
@@ -3,7 +3,7 @@
 #include <map>
 #include <cassert>
 
-#include "pegr/script/ScriptHelper.hpp"
+#include "pegr/script/Script_Helper.hpp"
 #include "pegr/script/Script.hpp"
 #include "pegr/logger/Logger.hpp"
