+++ src/pegr/test/Script_Test.cpp
@@ -1,10 +1,10 @@
 #include <stdexcept>
 #include <vector>
 
-#include "pegr/script/ScriptHelper.hpp"
+#include "pegr/script/Script_Helper.hpp"
 #include "pegr/script/Script.hpp"
 #include "pegr/logger/Logger.hpp"
-#include "pegr/test/TestUtil.hpp"
+#include "pegr/test/Test_Util.hpp"
 
 namespace pegr {
 namespace Test {
