+++ src/pegr/test/Unique_Ptr_Test.cpp
@@ -2,7 +2,7 @@
 #include <algorithm>
 #include <memory>
 
-#include "pegr/test/TestUtil.hpp"
+#include "pegr/test/Test_Util.hpp"
 
 namespace pegr {
 namespace Test {
